+++ cloudant-http/src/main/java/com/cloudant/http/internal/interceptors/IamCookieInterceptor.java
@@ -44,7 +44,7 @@ public class IamCookieInterceptor extends CookieInterceptorBase {
         // Read iamServer from system property, or set default
         try {
             this.iamServerUrl = new URL(System.getProperty("com.cloudant.client.iamserver",
-                    "https://iam.bluemix.net/identity/token"));
+                    "https://iam.cloud.ibm.com/identity/token"));
         } catch (MalformedURLException mue) {
             throw new RuntimeException("IAM server property was not a valid URL", mue);
         }