+++ CHANGES.md
@@ -1,5 +1,6 @@
 # UNRELEASED
 - [NEW] Added option for client to authenticate with IAM token server.
+- [FIXED] Updated the default IAM token server URL.
 
 # 2.14.0 (2018-12-18)
 - [NEW] Add `purge_seq` getter on `DbInfo` class that returns sequence value as `String`.